lint

+++ backend/src/repository/user.py
@@ -1,9 +1,6 @@
-from sqlalchemy import update
-
-from sqlalchemy import select, cast, Integer
+from sqlalchemy import select, update
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from src.database import Model
 from src.database.user import User
 
 
@@ -22,6 +19,6 @@ class UserRepository:
         return user
 
     @staticmethod
-    async def subtract_money(user_id: int, cost: int, session: AsyncSession) -> None:
+    async def subtract_money(user_id: int, cost: float, session: AsyncSession) -> None:
         await session.execute(update(User).where(User.id == user_id).values(balance=User.balance - cost))
         await session.commit()
